refactor: update route imports to use centralized type definitions
