+++ app/routes/admin+/$company.settings/forms/company-settings.action.server.ts
@@ -1,8 +1,8 @@
 import { parseWithZod } from '@conform-to/zod'
 import { dataWithSuccess } from 'remix-toast'
+import type { Route } from '../+types/route'
 import { updateCompany } from '../functions/mutations.server'
 import { INTENTS, companySettingsSchema as schema } from '../types'
-import type { Route } from './+types'
 
 export const action = async ({ request, params }: Route.ActionArgs) => {
   const submission = await parseWithZod(await request.formData(), { schema })
