+++ app/routes/admin+/$company.settings/forms/delete-company.action.server.ts
@@ -1,9 +1,9 @@
 import { parseWithZod } from '@conform-to/zod'
 import { redirect } from 'react-router'
 import { $path } from 'safe-routes'
+import type { Route } from '../+types/route'
 import { deleteCompany } from '../functions.server'
 import { INTENTS, deleteCompanySchema as schema } from '../types'
-import type { Route } from './+types'
 
 export const action = async ({ request, params }: Route.ActionArgs) => {
   const submission = parseWithZod(await request.formData(), { schema })
